@@ -22,21 +22,21 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.13",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "^7.2.13",
-    "@angular/compiler": "^7.2.13",
-    "@angular/core": "^7.2.13",
-    "@angular/flex-layout": "^7.0.0-beta.24",
-    "@angular/forms": "^7.2.13",
-    "@angular/http": "^7.2.13",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "^7.2.13",
-    "@angular/platform-browser-dynamic": "^7.2.13",
-    "@angular/router": "^7.2.13",
-    "@caiu/library": "^7.1.31",
-    "@ngrx/effects": "^7.4.0",
-    "@ngrx/store": "^7.4.0",
+    "@angular/animations": "~8.0.0",
+    "@angular/cdk": "^8.0.1",
+    "@angular/common": "~8.0.0",
+    "@angular/compiler": "~8.0.0",
+    "@angular/core": "~8.0.0",
+    "@angular/flex-layout": "^8.0.0-beta.26",
+    "@angular/forms": "~8.0.0",
+    "@angular/http": "^7.2.15",
+    "@angular/material": "^8.0.1",
+    "@angular/platform-browser": "~8.0.0",
+    "@angular/platform-browser-dynamic": "~8.0.0",
+    "@angular/router": "~8.0.0",
+    "@caiu/library": "^7.2.30",
+    "@ngrx/effects": "^8.0.1",
+    "@ngrx/store": "^8.0.1",
     "@progress/kendo-angular-buttons": "^4.3.2",
     "@progress/kendo-angular-dateinputs": "^3.5.1",
     "@progress/kendo-angular-dropdowns": "^3.4.0",
@@ -47,36 +47,36 @@
     "@progress/kendo-angular-l10n": "^1.3.0",
     "@progress/kendo-data-query": "^1.5.0",
     "@progress/kendo-drawing": "^1.5.7",
-    "core-js": "^2.5.7",
     "hammerjs": "^2.0.8",
-    "rxjs": "~6.3.3",
+    "rxjs": "~6.4.0",
     "rxjs-compat": "^6.3.3",
     "tinymce": "^4.9.0",
-    "zone.js": "^0.8.26"
+    "tslib": "^1.9.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-ng-packagr": "~0.11.0",
-    "ng-packagr": "~4.4.0",
-    "tsickle": ">=0.34.0",
-    "tslib": "^1.9.3",
-    "@angular-devkit/build-angular": "~0.11.0",
-    "@angular/cli": "~7.1.3",
-    "@angular/compiler-cli": "^7.2.13",
-    "@angular/language-service": "^7.2.13",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
+    "@angular-devkit/build-ng-packagr": "~0.800.2",
+    "ng-packagr": "~5.3.0",
+    "tsickle": ">=0.35.0",
+    "tslib": "^1.10.0",
+    "@angular-devkit/build-angular": "~0.800.0",
+    "@angular/cli": "~8.0.2",
+    "@angular/compiler-cli": "~8.0.0",
+    "@angular/language-service": "~8.0.0",
     "@types/node": "~8.9.4",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~3.3.0",
+    "@types/jasmine": "~3.3.8",
+    "@types/jasminewd2": "~2.0.3",
+    "codelyzer": "^5.0.0",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.1.1",
+    "karma": "~4.1.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.4",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
     "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "^1.4.0",
-    "protractor": "~5.4.1",
+    "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.2.0"
+    "tslint": "~5.15.0",
+    "typescript": "~3.4.3"
   }
 }
