@@ -1,14 +1,14 @@
 {
   "name": "@caiu/library",
-  "version": "7.2.7",
+  "version": "8.0.0",
   "license": "MIT",
   "private": false,
   "peerDependencies": {
-    "@angular/common": "^7.3.7",
-    "@angular/core": "7.3.7",
-    "@angular/cdk": "^7.3.7",
-    "@angular/flex-layout": "^7.0.0-beta.24",
-    "@ngrx/effects": "^7.4.0",
-    "@ngrx/store": "^7.4.0"
+    "@angular/common": "~8.0.0",
+    "@angular/core": "~8.0.0",
+    "@angular/cdk": "^8.0.1",
+    "@angular/flex-layout": "^8.0.0-beta.26",
+    "@ngrx/effects": "^8.0.1",
+    "@ngrx/store": "^8.0.1"
   }
 }
