@@ -7,13 +7,19 @@ import { truthy } from '../../shared/utils';
   name: 'address'
 })
 export class AddressPipe implements PipeTransform {
-
   transform(value: Address, showName = false): string {
-    const address = `
+    const address = value
+      ? `
       ${value.address1 || ''}
-      ${value.city || ''}${truthy(value.city) && truthy(value.stateCode) ? ', ' + value.stateCode : ''} ${value.zip || ''}
-    `;
-    return showName ? `${value.firstName || ''} ${value.lastName || ''} ${address}` : address;
+      ${value.city || ''}${
+          truthy(value.city) && truthy(value.stateCode)
+            ? ', ' + value.stateCode
+            : ''
+        } ${value.zip || ''}
+    `
+      : '';
+    return showName
+      ? `${value.firstName || ''} ${value.lastName || ''} ${address}`
+      : address;
   }
-
 }
